Фикс на Double в новых функциях фрейма

+++ Sources/GXUtilz/Common/Display.swift
@@ -26,14 +26,14 @@ public extension CGFloat {
     /// Функция для взятия фрейма от ширины экрана в процентах
     /// - Parameter percent: Процент от ширины
     /// - Returns: Ширина фрейма
-    static func displayWidth(_ percent: Int) -> CGFloat {
+    static func displayWidth(_ percent: Double) -> CGFloat {
         Display.width * CGFloat(percent) / 100
     }
     
     /// Функция для взятия фрейма от высоты экрана в процентах
     /// - Parameter percent: Процент от высоты
     /// - Returns: Высота фрейма
-    static func displayHeight(_ percent: Int) -> CGFloat {
+    static func displayHeight(_ percent: Double) -> CGFloat {
         Display.height * CGFloat(percent) / 100
     }
 }
